backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lexicounsel-backend"
version = "0.1.0"
description = "Production-oriented legal consultation assistant backend"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "beautifulsoup4>=4.12.3",
  "faiss-cpu>=1.9.0",
  "fastapi>=0.115.6",
  "greenlet>=3.1.1",
  "httpx>=0.28.1",
  "langchain>=0.3.13",
  "langchain-community>=0.3.13",
  "langchain-openai>=0.2.14",
  "langchain-text-splitters>=0.3.4",
  "langsmith>=0.2.6",
  "orjson>=3.10.12",
  "pydantic-settings>=2.7.0",
  "pypdf>=5.1.0",
  "python-json-logger>=2.0.7",
  "python-multipart>=0.0.20",
  "rank-bm25>=0.2.2",
  "sqlalchemy>=2.0.36",
  "tenacity>=9.0.0",
  "tiktoken>=0.8.0",
  "slowapi>=0.1.9",
  "uvicorn[standard]>=0.34.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.0.0",
  "stripe>=10.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.4"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
